Only try adding a new element once
When no constraint is present which can cause certain elements to be
invalid when grouped together but not other elements, then there is no
reason to attempt to add different elements in place of another. If one
element cannot be added, then no other element can be added either, so
there's no reason to try.